L017-current user & added protection for all requests

+++ src/main/java/com/foodtech/blog/file/service/FileApiService.java
@@ -1,14 +1,14 @@
 package com.foodtech.blog.file.service;
 
+import com.foodtech.blog.auth.exceptions.AuthException;
+import com.foodtech.blog.auth.exceptions.NotAccessException;
+import com.foodtech.blog.auth.service.AuthService;
 import com.foodtech.blog.base.api.request.SearchRequest;
 import com.foodtech.blog.base.api.response.SearchResponse;;
-import com.foodtech.blog.file.mapping.FileMapping;
-import com.foodtech.blog.file.api.response.FileResponse;
-import com.foodtech.blog.file.exeception.FileExistException;
-import com.foodtech.blog.file.exeception.FileNotExistException;
+import com.foodtech.blog.base.servise.CheckAccess;
 import com.foodtech.blog.file.model.FileDoc;
 import com.foodtech.blog.file.repository.FileRepository;
-import com.foodtech.blog.user.exeception.UserNotExistException;
+import com.foodtech.blog.base.api.model.UserDoc;
 import com.foodtech.blog.user.repository.UserRepository;
 import com.mongodb.BasicDBObject;
 import com.mongodb.DBObject;
@@ -21,13 +21,9 @@ import org.springframework.data.mongodb.core.query.Criteria;
 import org.springframework.data.mongodb.core.query.Query;
 import org.springframework.data.mongodb.gridfs.GridFsOperations;
 import org.springframework.data.mongodb.gridfs.GridFsTemplate;
-import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Service;
-import org.springframework.util.DigestUtils;
-import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.multipart.MultipartFile;
 
-import javax.swing.text.ChangedCharSetException;
 import java.io.IOException;
 import java.io.InputStream;
 import java.util.List;
@@ -35,15 +31,16 @@ import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
-public class FileApiService {
+public class FileApiService extends CheckAccess<FileDoc> {
     private final FileRepository fileRepository;
     private final MongoTemplate mongoTemplate;
     private final GridFsTemplate gridFsTemplate;
     private final GridFsOperations operations;
     private final UserRepository userRepository;
+    private final AuthService authService;
 
-    public FileDoc create(MultipartFile file, ObjectId ownerId) throws  IOException, UserNotExistException {
-        if(userRepository.findById(ownerId).isEmpty()) throw new UserNotExistException();
+    public FileDoc create(MultipartFile file) throws IOException, AuthException {
+        UserDoc owner = authService.currentUser();
 
         DBObject metaData = new BasicDBObject();
         metaData.put("type",file.getContentType());
@@ -56,7 +53,7 @@ public class FileApiService {
         FileDoc fileDoc = FileDoc.builder()
                 .id(id)
                 .title(file.getOriginalFilename())
-                .ownerId(ownerId)
+                .ownerId(owner.getId())
                 .contentType(file.getContentType())
                 .build();
 
@@ -94,8 +91,19 @@ public class FileApiService {
         return SearchResponse.of(fileDocs, count);
     }
 
-    public void delete(ObjectId id){
+    public void delete(ObjectId id) throws NotAccessException, AuthException, ChangeSetPersister.NotFoundException {
+        checkAccess(fileRepository.findById(id).orElseThrow(ChangeSetPersister.NotFoundException::new));
         gridFsTemplate.delete(new Query(Criteria.where("_id").is(id)));
         fileRepository.deleteById(id);
+    }
+
+    @Override
+    protected ObjectId getOwnerFromEntity(FileDoc entity) {
+        return entity.getOwnerId();
+    }
+
+    @Override
+    protected AuthService authService() {
+        return authService;
     }
 }
