+++ src/main/java/com/foodtech/blog/photo/service/PhotoApiService.java
@@ -1,19 +1,20 @@
 package com.foodtech.blog.photo.service;
 
 import com.foodtech.blog.album.exeception.AlbumNotExistException;
+import com.foodtech.blog.album.model.AlbumDoc;
 import com.foodtech.blog.album.repository.AlbumRepository;
-import com.foodtech.blog.base.api.request.SearchRequest;
+import com.foodtech.blog.auth.exceptions.AuthException;
+import com.foodtech.blog.auth.exceptions.NotAccessException;
+import com.foodtech.blog.auth.service.AuthService;
 import com.foodtech.blog.base.api.response.SearchResponse;
+import com.foodtech.blog.base.servise.CheckAccess;
 import com.foodtech.blog.photo.api.request.PhotoRequest;
 import com.foodtech.blog.photo.api.request.PhotoSearchRequest;
 import com.foodtech.blog.photo.mapping.PhotoMapping;
-import com.foodtech.blog.photo.api.response.PhotoResponse;
-import com.foodtech.blog.photo.exeception.PhotoExistException;
 import com.foodtech.blog.photo.exeception.PhotoNotExistException;
 import com.foodtech.blog.photo.model.PhotoDoc;
 import com.foodtech.blog.photo.repository.PhotoRepository;
-import com.foodtech.blog.user.exeception.UserNotExistException;
-import com.foodtech.blog.user.repository.UserRepository;
+import com.foodtech.blog.base.api.model.UserDoc;
 import com.mongodb.BasicDBObject;
 import com.mongodb.DBObject;
 import com.mongodb.client.gridfs.model.GridFSFile;
@@ -25,10 +26,7 @@ import org.springframework.data.mongodb.core.query.Criteria;
 import org.springframework.data.mongodb.core.query.Query;
 import org.springframework.data.mongodb.gridfs.GridFsOperations;
 import org.springframework.data.mongodb.gridfs.GridFsTemplate;
-import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Service;
-import org.springframework.util.DigestUtils;
-import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.IOException;
@@ -38,17 +36,19 @@ import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
-public class PhotoApiService {
+public class PhotoApiService extends CheckAccess<PhotoDoc> {
     private final PhotoRepository photoRepository;
     private final MongoTemplate mongoTemplate;
     private final AlbumRepository albumRepository;
-    private final UserRepository userRepository;
     private final GridFsTemplate gridFsTemplate;
     private final GridFsOperations operations;
+    private final AuthService authService;
 
-    public PhotoDoc create(MultipartFile file, ObjectId ownerId, ObjectId albumId) throws AlbumNotExistException, UserNotExistException, IOException {
-        if(albumRepository.findById(albumId).isEmpty()) throw new AlbumNotExistException();
-        if(userRepository.findById(ownerId).isEmpty()) throw new UserNotExistException();
+    public PhotoDoc create(MultipartFile file, ObjectId albumId) throws AlbumNotExistException, IOException, AuthException, NotAccessException {
+        UserDoc userDoc = authService.currentUser();
+        AlbumDoc albumDoc =albumRepository.findById(albumId).orElseThrow(AlbumNotExistException::new);
+
+        if(albumDoc.getOwnerId().equals(userDoc.getId()) == false) throw new NotAccessException();
 
         DBObject metaData = new BasicDBObject();
         metaData.put("type",file.getContentType());
@@ -62,7 +62,7 @@ public class PhotoApiService {
                 .id(id)
                 .albumId(albumId)
                 .title(file.getOriginalFilename())
-                .ownerId(ownerId)
+                .ownerId(userDoc.getId())
                 .contentType(file.getContentType())
                 .build();
         photoRepository.save(photoDoc);
@@ -98,13 +98,14 @@ public class PhotoApiService {
         return SearchResponse.of(photoDocs, count);
     }
 
-    public PhotoDoc update(PhotoRequest request) throws PhotoNotExistException {
+    public PhotoDoc update(PhotoRequest request) throws PhotoNotExistException, NotAccessException, AuthException {
         Optional<PhotoDoc> photoDocOptional = photoRepository.findById(request.getId());
         if(photoDocOptional.isPresent() == false){
             throw new PhotoNotExistException();
         }
         PhotoDoc oldDoc = photoDocOptional.get();
-        PhotoDoc photoDoc = PhotoMapping.getInstance().getRequest().convert(request);
+        UserDoc owner = checkAccess(oldDoc);
+        PhotoDoc photoDoc = PhotoMapping.getInstance().getRequest().convert(request, owner.getId());
 
         photoDoc.setId(request.getId());
         photoDoc.setAlbumId(oldDoc.getAlbumId());
@@ -115,8 +116,19 @@ public class PhotoApiService {
         return photoDoc;
     }
 
-    public void delete(ObjectId id){
+    public void delete(ObjectId id) throws NotAccessException, AuthException, ChangeSetPersister.NotFoundException {
+        checkAccess(photoRepository.findById(id).orElseThrow(ChangeSetPersister.NotFoundException::new));
         gridFsTemplate.delete(new Query(Criteria.where("_id").is(id)));
         photoRepository.deleteById(id);
+    }
+
+    @Override
+    protected ObjectId getOwnerFromEntity(PhotoDoc entity) {
+        return entity.getOwnerId();
+    }
+
+    @Override
+    protected AuthService authService() {
+        return authService;
     }
 }
